Move input bindings to separated package
