+++ core/frontend/src/main/scala/io/udash/bindings/inputs/CheckButtons.scala
@@ -1,9 +1,8 @@
-package io.udash.bindings
+package io.udash.bindings.inputs
 
 import io.udash.properties.seq.SeqProperty
 import io.udash.properties.single.ReadableProperty
 import org.scalajs.dom.{html, _}
-
 import scalatags.JsDom
 import scalatags.JsDom.all._
