+++ core/frontend/src/main/scala/io/udash/bindings/inputs/Checkbox.scala
@@ -1,8 +1,7 @@
-package io.udash.bindings
+package io.udash.bindings.inputs
 
 import io.udash.properties.single.Property
 import org.scalajs.dom.{html, _}
-
 import scalatags.JsDom
 import scalatags.JsDom.all._
