+++ core/frontend/src/main/scala/io/udash/bindings/inputs/TextArea.scala
@@ -1,45 +1,44 @@
-package io.udash.bindings
+package io.udash.bindings.inputs
 
-import io.udash.bindings.modifiers.TextInputsModifier
 import io.udash.properties.single.Property
 import org.scalajs.dom.{html, _}
-
-import scala.concurrent.duration.{Duration, DurationInt}
 import scalatags.JsDom
 import scalatags.JsDom.all._
 
+import scala.concurrent.duration.{Duration, DurationInt}
+
 /**
-  * Abstraction for HTML input tags.
+  * Simple HTML text area with bound Property.
   */
-private[bindings] abstract class Input(inputType: String) {
+object TextArea {
   /**
     * @param property Property to bind.
     * @param debounce Property update timeout after input changes.
     * @param xs Additional Modifiers, don't use modifiers on value, onchange and onkeyup attributes.
     * @return HTML textarea with bound Property, applied modifiers and nested options.
     */
-  def apply(property: Property[String], debounce: Option[Duration], xs: Modifier*): JsDom.TypedTag[html.Input] = {
+  def apply(property: Property[String], debounce: Option[Duration], xs: Modifier*): JsDom.TypedTag[html.TextArea] = {
     val bind = new TextInputsModifier(property, debounce) {
       override def elementValue(t: Element): String =
-        t.asInstanceOf[html.Input].value
+        t.asInstanceOf[html.TextArea].value
 
       override def setElementValue(t: Element, v: String): Unit =
-        t.asInstanceOf[html.Input].value = if (v != null) v else ""
+        t.asInstanceOf[html.TextArea].value = v
 
       override def setElementKeyUp(t: Element, callback: (KeyboardEvent) => Any): Unit =
-        t.asInstanceOf[html.Input].onkeyup = callback
+        t.asInstanceOf[html.TextArea].onkeyup = callback
 
       override def setElementOnChange(t: Element, callback: (Event) => Any): Unit =
-        t.asInstanceOf[html.Input].onchange = callback
+        t.asInstanceOf[html.TextArea].onchange = callback
 
       override def setElementOnInput(t: Element, callback: (Event) => Any): Unit =
-        t.asInstanceOf[html.Input].oninput = callback
+        t.asInstanceOf[html.TextArea].oninput = callback
 
       override def setElementOnPaste(t: Element, callback: (Event) => Any): Unit =
-        t.asInstanceOf[html.Input].onpaste = callback
+        t.asInstanceOf[html.TextArea].onpaste = callback
     }
 
-    input(tpe := inputType, bind, xs)
+    textarea(bind, xs)
   }
 
   /**
@@ -48,7 +47,7 @@ private[bindings] abstract class Input(inputType: String) {
     * @return HTML textarea with bound Property, applied modifiers and nested options.
     */
   @deprecated(message = "You should use `debounced` method or explicitly pass `None` as debouncing parameter.", since = "0.3.0")
-  def apply(property: Property[String], xs: Modifier*): JsDom.TypedTag[html.Input] =
+  def apply(property: Property[String], xs: Modifier*): JsDom.TypedTag[html.TextArea] =
     apply(property, None, xs:_*)
 
   /**
@@ -56,15 +55,6 @@ private[bindings] abstract class Input(inputType: String) {
     * @param xs Additional Modifiers, don't use modifiers on value, onchange and onkeyup attributes.
     * @return HTML textarea with bound Property, applied modifiers and nested options.
     */
-  def debounced(property: Property[String], xs: Modifier*): JsDom.TypedTag[html.Input] =
+  def debounced(property: Property[String], xs: Modifier*): JsDom.TypedTag[html.TextArea] =
     apply(property, Some(20 millis), xs:_*)
 }
-
-/** Simple text input. */
-object TextInput extends Input("text")
-
-/** Password text input. */
-object PasswordInput extends Input("password")
-
-/** Number input. (HTML5) */
-object NumberInput extends Input("number")
